refactor(redux): 상위에서 state를 가져온다
주소창에서 바로 장바구니로 넘어간 경우 ProductList를 거쳐가지 않았기 때문에 상품 정보가 없다. 그래서 상위에서 상품 정보를 가져온다
